+++ Lib/fontbakery/profiles/shared_conditions.py
@@ -117,11 +117,12 @@ def ligatures(ttFont):
           for index in record.Feature.LookupListIndex:
             lookup = ttFont["GSUB"].table.LookupList.Lookup[index]
             for subtable in lookup.SubTable:
-              for firstGlyph in subtable.ligatures.keys():
-                all_ligatures[firstGlyph] = []
-                for lig in subtable.ligatures[firstGlyph]:
-                  if lig.Component not in all_ligatures[firstGlyph]:
-                    all_ligatures[firstGlyph].append(lig.Component)
+              if subtable.Format == 1:
+                for firstGlyph in subtable.ligatures.keys():
+                  all_ligatures[firstGlyph] = []
+                  for lig in subtable.ligatures[firstGlyph]:
+                    if lig.Component not in all_ligatures[firstGlyph]:
+                      all_ligatures[firstGlyph].append(lig.Component)
     return all_ligatures
   except:
     return -1 # Indicate fontTools-related crash...
@@ -137,10 +138,11 @@ def ligature_glyphs(ttFont):
           for index in record.Feature.LookupListIndex:
             lookup = ttFont["GSUB"].table.LookupList.Lookup[index]
             for subtable in lookup.SubTable:
-              for firstGlyph in subtable.ligatures.keys():
-                for lig in subtable.ligatures[firstGlyph]:
-                  if lig.LigGlyph not in all_ligature_glyphs:
-                    all_ligature_glyphs.append(lig.LigGlyph)
+              if subtable.Format == 1:
+                for firstGlyph in subtable.ligatures.keys():
+                  for lig in subtable.ligatures[firstGlyph]:
+                    if lig.LigGlyph not in all_ligature_glyphs:
+                      all_ligature_glyphs.append(lig.LigGlyph)
     return all_ligature_glyphs
   except:
     return -1  # Indicate fontTools-related crash...
@@ -345,3 +347,29 @@ def is_cjk_font(ttFont):
 @condition
 def typo_metrics_enabled(ttFont):
   return ttFont['OS/2'].fsSelection & 0b10000000 > 0
+
+
+@condition
+def is_indic_font(ttFont):
+  INDIC_FONT_DETECTION_CODEPOINTS = [
+    0x0988, # Bengali
+    0x0908, # Devanagari
+    0x0A88, # Gujarati
+    0x0A08, # Gurmukhi
+    0x0D08, # Kannada
+    0x0B08, # Malayalam
+    0xABC8, # Meetei Mayek
+    0x1C58, # OlChiki
+    0x0B08, # Oriya
+    0x0B88, # Tamil
+    0x0C08, # Telugu
+  ]
+
+  font_codepoints = ttFont['cmap'].getBestCmap().keys()
+  for codepoint in INDIC_FONT_DETECTION_CODEPOINTS:
+    if codepoint in font_codepoints:
+      return True
+
+  #otherwise:
+  return False
+